+++ web/air/src/components/utils.js
@@ -24,6 +24,6 @@ export async function onLarkOAuthClicked(lark_client_id) {
   if (!state) return;
   let redirect_uri = `${window.location.origin}/oauth/lark`;
   window.open(
-    `https://open.feishu.cn/open-apis/authen/v1/index?redirect_uri=${redirect_uri}&app_id=${lark_client_id}&state=${state}`
+    `https://accounts.feishu.cn/open-apis/authen/v1/authorize?client_id=${lark_client_id}&response_type=code&redirect_uri=${redirect_uri}&state=${state}`
   );
 }